build: add linux build tags and relocate portable aetom-gvisor code to internal package
